+++ Demo/src/main/java/com/paypal/android/api/services/SDKSampleServerAPI.kt
@@ -13,7 +13,6 @@ import com.paypal.checkout.order.OrderRequest
 import okhttp3.OkHttpClient
 import okhttp3.ResponseBody
 import okhttp3.logging.HttpLoggingInterceptor
-import org.json.JSONArray
 import org.json.JSONObject
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
@@ -255,13 +254,11 @@ class SDKSampleServerAPI {
 
         val responseJSON = JSONObject(response.string())
         val customerJSON = responseJSON.getJSONObject("customer")
-        val approveVaultHref = findApprovalHref(responseJSON)
 
         PayPalSetupToken(
             id = responseJSON.getString("id"),
             customerId = customerJSON.getString("id"),
-            status = responseJSON.getString("status"),
-            approveVaultHref = approveVaultHref
+            status = responseJSON.getString("status")
         )
     }
 
@@ -289,16 +286,5 @@ class SDKSampleServerAPI {
             vaultId = optNonEmptyString(vaultJSON, "id"),
             customerId = optNonEmptyString(vaultCustomerJSON, "id")
         )
-    }
-
-    private fun findApprovalHref(responseJSON: JSONObject): String? {
-        val linksJSON = responseJSON.optJSONArray("links") ?: JSONArray()
-        for (i in 0 until linksJSON.length()) {
-            val link = linksJSON.getJSONObject(i)
-            if (link.getString("rel") == "approve") {
-                return link.getString("href")
-            }
-        }
-        return null
     }
 }