+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # PayPal Android SDK Release Notes
 
+## unreleased
+* PayPalWebPayments
+  * Deprecate `PayPalWebVaultRequest(setupTokenId, approveVaultHref)`
+  * Add `PayPalWebVaultRequest(setupTokenId)`
+
 ## 1.4.1 (2024-06-06)
 * PaymentButtons
   * Add `paylater_color` to `PayLaterButton` to control the color of the pay later button from XML.