+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # PayPal Android SDK Release Notes
 
+## unreleased
+* PayPalNativePayments (DEPRECATED)
+  * **NOTE:** This module is being deprecated and will be removed in the future version of the SDK
+  * Add deprecated warning message to all classes
+  
 ## 1.4.1 (2024-06-06)
 * PaymentButtons
   * Add `paylater_color` to `PayLaterButton` to control the color of the pay later button from XML.