+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## unreleased
 * PayPalNativePayments (DEPRECATED)
-  * **NOTE:** This module is being deprecated and will be removed in the future version of the SDK
+  * **NOTE:** This module is being deprecated and will be removed in the future version of the SDK. Use `PayPalWebPayments` module instead.
   * Add deprecated warning message to all classes
   
 ## 1.4.1 (2024-06-06)